pyproject.toml
[tool.poetry]
name = "discord-archiver"
version = "0.1.1"
description = ""
authors = ["Taylor Kettle <tin.teapot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
PyInquirer = "^1.0.3"
pathvalidate = "^2.4.1"
pathos = "^0.2.8"
pyinstaller = "^4.5.1"
python-dotenv = "^0.19.0"
requests-futures = "^1.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
